+++ AddOns/PitBull4/Modules/Aura/Options.lua
@@ -273,6 +273,7 @@ PitBull4_Aura:SetDefaults({
 		-- % Race map filters
 		-- & Class map filters
 		-- * Extra filters
+		-- ++ Evoker
 		-- + Death Knight
 		-- ,, Demon Hunter
 		-- , Druid
@@ -292,6 +293,7 @@ PitBull4_Aura:SetDefaults({
 		-- 7 Night Elf
 		-- 77 Void Elf
 		-- 8 Gnome
+		-- 88 Mechagnome
 		-- 9 Draenei
 		-- 99 Lightforged Draenei
 		-- : Orc
@@ -305,6 +307,7 @@ PitBull4_Aura:SetDefaults({
 		-- >> Nightborne
 		-- ? Worgen
 		-- ?? Goblin
+		-- @@ Dracthyr
 		-- @ Pandaren (shared, race filters are always followed by a number)
 		-- @ Simple filters
 		--
@@ -494,6 +497,13 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Should show',
 			should_show = true,
 			disabled = true,
+			built_in = true,
+		},
+		['++3'] = {
+			display_name = L["%s can dispel"]:format(LN.EVOKER),
+			filter_type = 'Aura Type',
+			whitelist = true,
+			aura_type_list = can_dispel.EVOKER,
 			built_in = true,
 		},
 		[',3'] = {
@@ -1119,6 +1129,13 @@ PitBull4_Aura:SetDefaults({
 			name_list = self_debuffs.WARRIOR,
 			built_in = true,
 		},
+		['++7'] = {
+			display_name = L["%s can purge"]:format(LN.EVOKER),
+			filter_type = 'Aura Type',
+			whitelist = true,
+			aura_type_list = can_purge.EVOKER,
+			built_in = true,
+		},
 		['+7'] = {
 			display_name = L["%s can purge"]:format(LN.DEATHKNIGHT),
 			filter_type = 'Aura Type',
@@ -1208,6 +1225,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '++3',
 				['DEATHKNIGHT'] = '@J',
 				['DEMONHUNTER'] = '@J',
 				['DRUID'] = ',3',
@@ -1228,6 +1246,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '+0',
 				['DEMONHUNTER'] = ',,0',
 				['DRUID'] = ',0',
@@ -1248,6 +1267,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '+1',
 				['DEMONHUNTER'] = '@J',
 				['DRUID'] = '@J',
@@ -1268,6 +1288,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '+2',
 				['DEMONHUNTER'] = ',,2',
 				['DRUID'] = ',2',
@@ -1288,6 +1309,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '+6',
 				['DEMONHUNTER'] = ',,6',
 				['DRUID'] = ',6',
@@ -1308,6 +1330,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '@J',
 				['DEMONHUNTER'] = '@J',
 				['DRUID'] = '@J',
@@ -1328,6 +1351,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '@J',
 				['DEATHKNIGHT'] = '@J',
 				['DEMONHUNTER'] = '@J',
 				['DRUID'] = '@J',
@@ -1348,6 +1372,7 @@ PitBull4_Aura:SetDefaults({
 			filter_type = 'Map',
 			map_type = 'class',
 			map = {
+				['EVOKER'] = '++7',
 				['DEATHKNIGHT'] = '+7',
 				['DEMONHUNTER'] = ',,7',
 				['DRUID'] = ',7',
@@ -1384,11 +1409,14 @@ PitBull4_Aura:SetDefaults({
 				['DarkIronDwarf'] = '660',
 				['LightforgedDraenei'] = '@J',
 				['VoidElf'] = '770',
-				-- ['KulTiranHuman'] = '@J',
 				['MagharOrc'] = '::0',
 				['HighmountainTauren'] = '@J',
 				['Nightborne'] = '@J',
+				-- ['KulTiranHuman'] = '@J',
 				-- ['ZandalariTroll'] = '@J',
+				-- ['Vulpera'] = '@J',
+				-- ['Mechagnome'] = '@J',
+				-- ['Dracthyr'] = '@J',
 			},
 			built_in = true,
 		},
@@ -1413,11 +1441,14 @@ PitBull4_Aura:SetDefaults({
 				['DarkIronDwarf'] = '662',
 				['LightforgedDraenei'] = '992',
 				['VoidElf'] = '772',
-				-- ['KulTiranHuman'] = '552',
 				['MagharOrc'] = '::2',
 				['HighmountainTauren'] = '<<2',
 				['Nightborne'] = '>>2',
+				-- ['KulTiranHuman'] = '552',
 				-- ['ZandalariTroll'] = '==2',
+				-- ['Vulpera'] = '@J',
+				-- ['Mechagnome'] = '@J',
+				-- ['Dracthyr'] = '@J',
 			},
 			built_in = true,
 		},
@@ -1442,11 +1473,14 @@ PitBull4_Aura:SetDefaults({
 				['DarkIronDwarf'] = '@J',
 				['LightforgedDraenei'] = '@J',
 				['VoidElf'] = '@J',
-				-- ['KulTiranHuman'] = '@J',
 				['MagharOrc'] = '@J',
 				['HighmountainTauren'] = '<<6',
 				['Nightborne'] = '>>6',
+				-- ['KulTiranHuman'] = '@J',
 				-- ['ZandalariTroll'] = '@J',
+				-- ['Vulpera'] = '@J',
+				-- ['Mechagnome'] = '@J',
+				-- ['Dracthyr'] = '@J',
 			},
 			built_in = true,
 		},
@@ -1471,11 +1505,14 @@ PitBull4_Aura:SetDefaults({
 				['DarkIronDwarf'] = '@J',
 				['LightforgedDraenei'] = '@J',
 				['VoidElf'] = '@J',
-				-- ['KulTiranHuman'] = '@J',
 				['MagharOrc'] = '@J',
 				['HighmountainTauren'] = '@J',
 				['Nightborne'] = '@J',
+				-- ['KulTiranHuman'] = '@J',
 				-- ['ZandalariTroll'] = '@J',
+				-- ['Vulpera'] = '@J',
+				-- ['Mechagnome'] = '@J',
+				-- ['Dracthyr'] = '@J',
 			},
 			built_in = true,
 		},