+++ AddOns/VuhDo/VuhDoEventHandler.lua
@@ -632,9 +632,11 @@ function VUHDO_OnEvent(_, anEvent, anArg1, anArg2, anArg3, anArg4, anArg5, anArg
 		if VUHDO_RAID then VUHDO_readyCheckEnds(); end
 
 	elseif "CVAR_UPDATE" == anEvent then
-		VUHDO_IS_SFX_ENABLED = tonumber(GetCVar("Sound_EnableSFX")) == 1;
-		VUHDO_IS_SOUND_ERRORSPEECH_ENABLED = tonumber(GetCVar("Sound_EnableErrorSpeech")) == 1;
-		if VUHDO_VARIABLES_LOADED then VUHDO_reloadUI(false); end
+		-- Patch 10.0.0 makes setting CVars freeze the game client
+		-- FIXME: also there is some issue where this event fires before bouquets have been properly decompressed
+		VUHDO_IS_SFX_ENABLED = false; --tonumber(GetCVar("Sound_EnableSFX")) == 1;
+		VUHDO_IS_SOUND_ERRORSPEECH_ENABLED = false; --tonumber(GetCVar("Sound_EnableErrorSpeech")) == 1;
+		--if VUHDO_VARIABLES_LOADED then VUHDO_reloadUI(false); end
 
 	elseif "INSPECT_READY" == anEvent then
 		VUHDO_inspectLockRole();