+++ AddOns/VuhDo/VuhDoLocalization_de.lua
@@ -501,3 +501,6 @@ VUHDO_I18N_TRINKET_2 = "Trinket 2";
 
 -- 3.139
 VUHDO_I18N_EVOKERS = "Evokers";
+
+-- 3.143
+VUHDO_I18N_BUFFC_EARTH_SHIELD = "Earth Shield (Self)";