+++ AddOns/HandyNotes/CHANGELOG.md
@@ -1,6 +1,6 @@
 # HandyNotes
 
-## [v1.6.9](https://github.com/Nevcairiel/HandyNotes/tree/v1.6.9) (2022-10-25)
-[Full Changelog](https://github.com/Nevcairiel/HandyNotes/compare/v1.6.8...v1.6.9) [Previous Releases](https://github.com/Nevcairiel/HandyNotes/releases)
+## [v1.6.12](https://github.com/Nevcairiel/HandyNotes/tree/v1.6.12) (2022-11-07)
+[Full Changelog](https://github.com/Nevcairiel/HandyNotes/compare/v1.6.11...v1.6.12) [Previous Releases](https://github.com/Nevcairiel/HandyNotes/releases)
 
-- Update TOC for 10.0  
+- Add a fallback to release the capture frame when the modifiers change without an event