+++ AddOns/PitBull4/CHANGELOG.md
@@ -1,7 +1,20 @@
 # PitBull Unit Frames 4.0 [WRONG VERSION]
 
-## [v4.2.21](https://github.com/nebularg/PitBull4/tree/v4.2.21) (2022-10-26)
-[Full Changelog](https://github.com/nebularg/PitBull4/compare/v4.2.20...v4.2.21) [Previous Releases](https://github.com/nebularg/PitBull4/releases)
+## [v4.2.22](https://github.com/nebularg/PitBull4/tree/v4.2.22) (2022-11-03)
+[Full Changelog](https://github.com/nebularg/PitBull4/compare/v4.2.21...v4.2.22) [Previous Releases](https://github.com/nebularg/PitBull4/releases)
 
-- ReputationBar: Fix Friendship rep  
-- SoulShards: Update  
+- Aura: Update dispel and purge spells  
+- Aura: Add Evoker and missing racial filter stubs  
+- LuaTexts: Update class and race short names  
+- LuaTexts: Add Shorter for numbers < 10000  
+    Even though 1623 and 1.6k are the same number of characters and  
+    the non-abbreviated form conveys more information, people get  
+    all OCD about it.  
+- Background: Fix portrait 3D question mark model  
+- LuaTexts: Add reaction colors for unknown and tapped; and unknown class color  
+- Add fader for phased units (#33)  
+- Fix portaits not fading  
+- Reputation: Fixes and update to show paragon progress  
+- Portrait: Fix 3D question mark model  
+- Portrait: Remove Pirate Day outfits  
+- Portrait: Use SetModelByFileID