+++ AddOns/RaiderIO/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Raider.IO Mythic Plus, Raiding, and Recruitment
 
-## [v202210300600](https://github.com/RaiderIO/raiderio-addon/tree/v202210300600) (2022-10-30)
-[Full Changelog](https://github.com/RaiderIO/raiderio-addon/compare/v202210290600...v202210300600) [Previous Releases](https://github.com/RaiderIO/raiderio-addon/releases)
+## [v202211100600](https://github.com/RaiderIO/raiderio-addon/tree/v202211100600) (2022-11-10)
+[Full Changelog](https://github.com/RaiderIO/raiderio-addon/compare/v202211090600...v202211100600) [Previous Releases](https://github.com/RaiderIO/raiderio-addon/releases)
 
 - [Raider.IO] Database Refresh