+++ AddOns/TradeSkillMaster/ChangeLog.md
@@ -1,6 +1,6 @@
-## v4.11.63 Changes
+## v4.11.66 Changes
 
-* [Retail] Fixed various issues with default / TSM frames not closing properly
-* [Retail] Fixed issue with crafting multiple of an item
+* [Retail] Fixed Destroy Next not working
+* [Retail] Fixed conflict with bag addons
 
 [Known Issues](https://support.tradeskillmaster.com/en_US/known_issues)